equipment name is has unique checker

+++ controller/equipment.controller.js
@@ -5,37 +5,39 @@ const shortid = require('shortid');
 const hbs = require('hbs');
 const { validationResult } = require('express-validator');
 
-hbs.registerHelper('subtract', function (a, b) { return a-b; });
+hbs.registerHelper('subtract', function (a, b) { return a - b; });
 
-exports.createEquipment = function (req, res) {
+exports.createEquipment = async function (req, res) {
 
     var errors = validationResult(req);
 
     if (errors.isEmpty()) {
-        const tempPath = req.file.path;
-        const filename = shortid.generate() + '.png';
-        const filePath = path.join(__dirname, '/../public/uploads/equipment-images', filename);
-        const relativeFilePath = '/uploads/equipment-images/' + filename;
-    
-        fs.rename(tempPath, filePath, async function(err) {
-            if (err) {
-                console.log(err);
-                res.redirect('/manage-equipment/');
-            } else {
-                let equipment = new Equipment({
-                    name: req.body.name,
-                    quantity: parseInt(req.body.count),
-                    imageURL: relativeFilePath
-                });
-    
-                try {
-                    await equipment.save();
-                } catch(err) {
+        var sameEquipCt = await Equipment.countDocuments(
+            { name: req.body.name });
+        if (sameEquipCt == 0) {
+            const tempPath = req.file.path;
+            const filename = shortid.generate() + '.png';
+            const filePath = path.join(__dirname, '/../public/uploads/equipment-images', filename);
+            const relativeFilePath = '/uploads/equipment-images/' + filename;
+            fs.rename(tempPath, filePath, async function (err) {
+                if (err) {
                     console.log(err);
+                } else {
+                    let equipment = new Equipment({
+                        name: req.body.name,
+                        quantity: parseInt(req.body.count),
+                        imageURL: relativeFilePath
+                    });
+
+                    try {
+                        await equipment.save();
+                    } catch (err) {
+                        console.log(err);
+                    }
                 }
-                res.redirect("/manage-equipment/");
-            }
-        });
+            });
+        }
+        res.redirect("/manage-equipment/");
     }
 };
 
@@ -60,44 +62,47 @@ exports.viewAllEquipment = async function (req, res) {
 exports.updateEquipment = async function (req, res) {
 
     var errors = validationResult(req);
-
     if (errors.isEmpty()) {
-        try {
-            var equipment = await Equipment.findById(req.body.equipmentid);
-            if (req.body.name.trim().length != 0) { equipment.name = req.body.name; }
-            if (!isNaN(parseInt(req.body.count))) { equipment.quantity = req.body.count; }
-            if (req.file != null) {
+        var sameEquipCt = await Equipment.countDocuments(
+            { name: req.body.name });
+        if (sameEquipCt == 0) {
+            try {
+                var equipment = await Equipment.findById(req.body.equipmentid);
+                if (req.body.name.trim().length != 0) { equipment.name = req.body.name; }
+                if (!isNaN(parseInt(req.body.count))) { equipment.quantity = req.body.count; }
+                if (req.file != null) {
     
-                if (fs.existsSync(path.join(__dirname, '/../public', equipment.imageURL)))
-                    fs.unlinkSync(path.join(__dirname, '/../public', equipment.imageURL));
+                    if (fs.existsSync(path.join(__dirname, '/../public', equipment.imageURL)))
+                        fs.unlinkSync(path.join(__dirname, '/../public', equipment.imageURL));
     
-                const tempPath = req.file.path;
-                const filename = shortid.generate() + '.png';
-                const filePath = path.join(__dirname, '/../public/uploads/equipment-images', filename);
-                const relativeFilePath = '/uploads/equipment-images/' + filename;
+                    const tempPath = req.file.path;
+                    const filename = shortid.generate() + '.png';
+                    const filePath = path.join(__dirname, '/../public/uploads/equipment-images', filename);
+                    const relativeFilePath = '/uploads/equipment-images/' + filename;
     
-                fs.rename(tempPath, filePath, async function(err) {
-                    if (err) console.log(err);
-                    else equipment.imageURL = relativeFilePath;
-                    try {
-                        await equipment.save();
-                    }
-                    catch (err) {
-                        console.log('Error updating db: ' + err);
-                    }
-                });
+                    fs.rename(tempPath, filePath, async function (err) {
+                        if (err) console.log(err);
+                        else equipment.imageURL = relativeFilePath;
+                        try {
+                            await equipment.save();
+                        }
+                        catch (err) {
+                            console.log('Error updating db: ' + err);
+                        }
+                    });
+                }
+                else {
+                    await equipment.save();
+                }
+            } catch (err) {
+                console.log('Error updating db: ' + err);
             }
-            else {
-                await equipment.save();
-            }
-        } catch (err) {
-            console.log('Error updating db: ' + err);
-        }              
+        }
         res.redirect("/manage-equipment/");
     }
 };
 
-exports.deleteEquipment = async function (req, res) { 
+exports.deleteEquipment = async function (req, res) {
     try {
         var equipment = await Equipment.findById(req.body.equipmentid);
         if (fs.existsSync(path.join(__dirname, '/../public', equipment.imageURL)))
@@ -105,7 +110,7 @@ exports.deleteEquipment = async function (req, res) {
         await Equipment.findByIdAndDelete(req.body.equipmentid);
     } catch (err) {
         console.log(err);
-    } 
+    }
     res.redirect("/manage-equipment/");
 };
